Adding Requested Featured: External Build Tools
Eclipse uses `*.launch` files stored in `.externalToolBuilders` folders inside cartridges in order to handle automagically building for things like Sass and JS Minification.

This Commit adds support during project setup to look for those setup scripts and add them to your project config.  Then during the watch process, if a file is being uploaded that was configured to trigger a build, it will do so.

If you have already setup a project, you will need to do so again with the same setting as last time to update the project config to use this new build setting.